Move some leftover functions to nodes.py, comments.py, strings.py
